Replaces `WithImportRegionalSingleton` with `WithImportByIdentity`
